ControllApp/requirements.txt
matplotlib
pygobject
pyserial